% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_freq.R
\name{gamfreq}
\alias{gamfreq}
\title{Gamete frequencies under a generalized model}
\usage{
gamfreq(
  g,
  ploidy,
  gam = NULL,
  alpha = NULL,
  beta = NULL,
  type = c("mix", "polysomic"),
  add_dr = TRUE
)
}
\arguments{
\item{g}{Parent genotype.}

\item{ploidy}{Parent ploidy. Should be even, and between 2 and 20 (inclusive).
Let me know if you need the ploidy to be higher. I can update
the package really easily.}

\item{gam}{The mixture proportions for the pairing configurations.
The proportions are in the same order the configurations in
\code{\link{seg}}. See Gerard et al (2018) for details on
pairing configurations.}

\item{alpha}{The double reduction rate(s) (if using). Defaults to 0's.}

\item{beta}{The double reduction adjustment for simplex markers if
\code{type = "mix"} and \code{add_dr = TRUE}. Assumed to be 0 by default.}

\item{type}{Either \code{"mix"}, meaning a mixture model of
pairing configurations, or \code{"polysomic"} for polysomic inheritance.}

\item{add_dr}{A logical. If \code{type = "polysomic"}, then the double
double reduction rate (\code{"alpha"}) will be used no matter the
value of \code{add_dr}, so set \code{alpha = 0} if you don't want it.
But if \code{type = "mix"} then we will incorporate double reduction
only in simplex markers (where it matters the most, and where
preferential pairing does not operate).}
}
\description{
Gamete frequencies under a generalized model
}
\references{
\itemize{
\item{Gerard, D., Ferrão, L. F. V., Garcia, A. A. F., & Stephens, M. (2018). Genotyping polyploids from messy sequencing data. Genetics, 210(3), 789-807.}
}
}
\author{
David Gerard
}
