% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{nextElem.arrayiter}
\alias{nextElem.arrayiter}
\title{Next element in an array}
\usage{
\method{nextElem}{arrayiter}(obj, ...)
}
\arguments{
\item{obj}{An arrayiter object}

\item{...}{not used}
}
\value{
The next sub-array.
}
\description{
This is applied to an \code{arrayiter} object to obtain the next sub-array
along one of the dimensions.
}
\examples{
glist <- multidog_to_g(
  mout = ufit,
  ploidy = 4,
  type = "all_gl",
  p1 = "indigocrisp",
  p2 = "sweetcrisp")
g <- iterators::iter(glist$g, by = 3)
head(iterators::nextElem(g))
head(iterators::nextElem(g))
head(iterators::nextElem(g))

}
\seealso{
\code{\link[=iter.array]{iter.array()}}
}
\author{
David Gerard
}
