% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscripts.R
\name{offspring_geno}
\alias{offspring_geno}
\title{Simulates genotypes given genotype frequencies.}
\usage{
offspring_geno(gf, n)
}
\arguments{
\item{gf}{Vector of offspring genotype frequencies}

\item{n}{Sample size}
}
\value{
Simulated genotypes
}
\description{
Takes as input the offspring
genotype frequencies and a sample size and returns simulated genotypes.
}
\examples{
set.seed(1)
gf <- offspring_gf_2(alpha = 1/6, xi1 = 1/3, xi2 = 1/3, p1 = 2, p2 = 3)
offspring_geno(gf = gf, n = 10)

}
\author{
Mira Thakkar
}
