% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscripts.R
\name{pvec_tet_3}
\alias{pvec_tet_3}
\title{Tetraploid gamete frequencies of gametes when one parent's genotype is known}
\usage{
pvec_tet_3(tau, beta, gamma, ell)
}
\arguments{
\item{tau}{Probability of quadrivalent formation}

\item{beta}{Probability of double reduction given quadrivalent formation}

\item{gamma}{Probability of AA/aa pairing given bivalent formation}

\item{ell}{The parent genotype}
}
\value{
The gamete genotype frequencies
}
\description{
This is under the three parameter model.
}
\examples{
pvec_tet_3(tau = 0.5, beta = 0.1, gamma = 0.5, ell = 2)

}
\author{
David Gerard
}
