% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segtest-package.R
\docType{package}
\name{segtest-package}
\alias{segtest}
\alias{segtest-package}
\title{segtest: Tests for Segregation Distortion in Polyploids}
\description{
Provides tests for segregation distortion in F1 polyploid populations under different assumptions of meiosis. These tests can account for double reduction, partial preferential pairing, and genotype uncertainty through the use of genotype likelihoods. Parallelization support is provided. Details of these methods are described in Gerard et al. (2025) \doi{10.1007/s00122-025-04816-z}. Part of this material is based upon work supported by the National Science Foundation under Grant No. 2132247. The opinions, findings, and conclusions or recommendations expressed are those of the author and do not necessarily reflect the views of the National Science Foundation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dcgerard.github.io/segtest/}
  \item Report bugs at \url{https://github.com/dcgerard/segtest/issues}
}

}
\author{
\strong{Maintainer}: David Gerard \email{gerard.1787@gmail.com} (\href{https://orcid.org/0000-0001-9450-5023}{ORCID})

Authors:
\itemize{
  \item Mira Thakkar \email{chrtmira@gmail.com}
}

Other contributors:
\itemize{
  \item NSF DBI 2132247 (https://www.nsf.gov/awardsearch/showAward?AWD_ID=2132247) [funder]
}

}
\keyword{internal}
