% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscripts.R
\name{three_to_two}
\alias{three_to_two}
\title{Convert from three parameters to two parameters}
\usage{
three_to_two(tau, beta, gamma)
}
\arguments{
\item{tau}{Probability of quadrivalent formation}

\item{beta}{Probability of double reduction given quadrivalent formation}

\item{gamma}{Probability of AA/aa pairing given bivalent formation}
}
\value{
A vector of length two. The first is the double reduction rate
(\code{alpha}), and the second is the preferential pairing parameter
(\code{xi}).
}
\description{
Convert from three parameters to two parameters
}
\examples{
three_to_two(tau = 0.1, beta = 1/6, gamma = 1/4)

}
\author{
David Gerard
}
