% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ufit}
\alias{ufit}
\alias{ufit2}
\alias{ufit3}
\title{Genotype data from Cappai et al. (2020)}
\format{
An object of type \code{multidog} output from \code{\link[updog]{multidog}()}.
\describe{
\item{\code{ufit}}{Uses the \code{model = "norm"} option.}
\item{\code{ufit2}}{Uses the \code{model = "f1pp"} option.}
\item{\code{ufit3}}{Uses the \code{model = "f1"} option.}
}

An object of class \code{multidog} of length 2.

An object of class \code{multidog} of length 2.
}
\source{
\doi{10.5281/zenodo.13715703}
}
\usage{
ufit

ufit2

ufit3
}
\description{
A subset of data from Cappai et al. (2020), fit using
\code{\link[updog]{multidog}()}. This just contains a random set of 10 loci.
}
\references{
\itemize{
\item{Cappai, F., Amadeu, R. R., Benevenuto, J., Cullen, R., Garcia, A., Grossman, A., Ferrão, L., & Munoz, P. (2020). High-resolution linkage map and QTL analyses of fruit firmness in autotetraploid blueberry. \emph{Frontiers in plant science}, 11, 562171. \doi{10.3389/fpls.2020.562171}.}
}
}
\keyword{datasets}
