% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{llike_li}
\alias{llike_li}
\title{Objective function for \code{\link{em_li}()}}
\usage{
llike_li(B, lpivec)
}
\arguments{
\item{B}{The log-likelihood matrix. Rows are individuals columns are
genotypes.}

\item{lpivec}{The log prior vector.}
}
\description{
Objective function for \code{\link{em_li}()}
}
\author{
David Gerard
}
