% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{multi_lrt}
\alias{multi_lrt}
\title{Parallelized likelihood ratio test for segregation distortion.}
\usage{
multi_lrt(
  g,
  p1,
  p2,
  drbound = 1/6,
  pp = TRUE,
  dr = TRUE,
  alpha = 0,
  xi1 = 1/3,
  xi2 = 1/3
)
}
\arguments{
\item{g}{One of two inputs
\itemize{
\item{A matrix of genotype counts. The rows index the locis and the columns index the genotypes.}
\item{An array of genotype log-likelihoods. The rows index the loci, the columns index the individuals, and the slices index the genotypes. Log-likelihoods are base e (natural log).}
}}

\item{p1}{One of two inputs
\itemize{
\item{A vector of parent 1's genotypes.}
\item{A matrix of parent 1's genotype log-likelihoods. The rows index the loci and the columns index the genotypes. Logs are in base e (natural log).}
}}

\item{p2}{One of two inputs
\itemize{
\item{A vector of parent 1's genotypes.}
\item{A matrix of parent 1's genotype log-likelihoods. The rows index the loci and the columns index the genotypes. Logs are in base e (natural log).}
}}

\item{drbound}{The upper bound on the double reduction rate.}

\item{pp}{Is (partial) preferential pairing possible (\code{TRUE}) or
not (\code{FALSE})?}

\item{dr}{Is double reduction possible (\code{TRUE}) or
not (\code{FALSE})?}

\item{alpha}{If \code{dr = FALSE}, this is the known rate of double
reduction.}

\item{xi1}{If \code{pp = FALSE}, this is the known preferential pairing
parameter of parent 1.}

\item{xi2}{If \code{pp = FALSE}, this is the known preferential pairing
parameter of parent 2.}
}
\description{
Uses the \code{future} package to implement parallelization support for
the likelihood ratio tests for segregation distortion. Right now, this is
only supported for tetraploids (allo, auto, or segemental).
}
\examples{
## Assuming genotypes are known (typically bad idea)
glist <- multidog_to_g(mout = ufit, type = "all_g", p1 = "indigocrisp", p2 = "sweetcrisp")
p1 <- glist$p1
p2 <- glist$p2
g <- glist$g

## Using genotype likelihoods (typically good idea)
glist <- multidog_to_g(mout = ufit, type = "all_gl", p1 = "indigocrisp", p2 = "sweetcrisp")
p1 <- glist$p1
p2 <- glist$p2
g <- glist$g


}
\author{
David Gerard
}
