% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{nextElem.arrayiter}
\alias{nextElem.arrayiter}
\title{Next element in an array}
\usage{
\method{nextElem}{arrayiter}(obj, ...)
}
\arguments{
\item{obj}{An arrayiter object}

\item{...}{not used}
}
\description{
Next element in an array
}
\seealso{
\code{\link[=iter.array]{iter.array()}}
}
\author{
David Gerard
}
