% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscripts.R
\name{po_gl}
\alias{po_gl}
\title{Generate genotype likelihoods from offspring genotypes.}
\usage{
po_gl(
  genovec,
  p1_geno,
  p2_geno,
  ploidy,
  seq = 0.01,
  rd = 10,
  bias = 1,
  od = 0.01
)
}
\arguments{
\item{genovec}{Offspring genotypes}

\item{p1_geno}{Parent 1 genotype}

\item{p2_geno}{Parent 2 genotype}

\item{ploidy}{Ploidy}

\item{seq}{Sequencing error rate}

\item{rd}{Read depth}

\item{bias}{Bias}

\item{od}{Overdispersion}
}
\value{
Genotype likelihoods
}
\description{
Takes as input (i) the parent genotypes,
(ii) the offspring genotype freq, (iii) sequencing error rate, (iv) read
depth, (v) bias, (vi) overdispersion and returns genotype likelihoods.
}
\author{
Mira Thakkar
}
